File-path finding is now more flexible.
* Both Corpora.py and Config.py both import files, however these files
were imported under the assumption that the os.getcwd() was the
directory in which the program was being executed. In contexts where
talktown was nested in a file hierarchy (ex: jukejoint), these files
would fail to load properly. Both files now import other files more
flexibly by appending the directory of the files themselves.