Added an event emitter object, lo_fi may use it.
* An event emitter object may now be passed to the Game object.
* If the event emitter object is set, enact_lo_fi_simulation() will use
it to emit the randomly selected TOTT events during town generation.
This means that listeners of that event emitter have the option of
using the names of these events as they are actually being generated!
* In order for many asynchronous libaries to work, a long running
process like lo_fi must sleep for some miniscule amount of time. One way
for TOTT to stay agnostic towards which asynchronous library is used,
TOTT implements the use of Python's standard libary Time. In order for
the asynchronous library to make use of Time, it must monkey patch it.
Many asynchronous libraries have this capability.